+++ src/entities/chat/index.ts
@@ -1,2 +1,2 @@
-export * from './ui';
-export * from './model';
+export { GameChat } from './ui/GameChat';
+export { MainChat } from './ui/MainChat';
