+++ src/entities/chat/model/useGameChat.ts
@@ -1,8 +1,9 @@
-import { useState, useEffect, useRef } from 'react';
 import type { KeyboardEvent } from 'react';
+import { useEffect, useRef, useState } from 'react';
+
+import { useAuthStore } from '@/features/auth'; // TODO: 오류!! useAuthStore 위치변경(shared)
 import { useSocket } from '@/shared/api/socket/socketContext';
 import type { ChatMessage } from '@/shared/model/types';
-import { useAuthStore } from '@/features/auth';
 
 export const useGameChat = () => {
   const { socket } = useSocket();
