+++ src/entities/chat/model/useMainChat.ts
@@ -1,8 +1,9 @@
-import { useState, useRef, useEffect } from 'react';
-import type { Friend } from '@/shared/model/types';
+import { useEffect, useRef, useState } from 'react';
+
+import { useAuthStore } from '@/features/auth'; // TODO: 오류!! useAuthStore 위치 변경 해야함(shared)
 import { mockChannels, mockFriends } from '@/mocks/chat.mock';
 import { useSocket } from '@/shared/api/socket/socketContext';
-import { useAuthStore } from '@/features/auth';
+import type { Friend } from '@/shared/model/types';
 
 export const useMainChat = () => {
   const { messages, sendMessage: emitMessage } = useSocket();
