+++ src/entities/chat/ui/ChannelSelect.tsx
@@ -1,4 +1,5 @@
 import { UserIcon } from '@heroicons/react/24/outline';
+
 import { cn } from '@/shared/lib/cn';
 
 interface ChannelSelectProps {
