+++ src/entities/chat/ui/ChatInput.tsx
@@ -1,4 +1,5 @@
 import type { ComponentProps } from 'react';
+
 import { cn } from '@/shared/lib/cn';
 
 type InputProps = Pick<
