+++ src/entities/chat/ui/ChatSendButton.tsx
@@ -1,4 +1,5 @@
 import { ArrowUpIcon } from '@heroicons/react/24/solid';
+
 import { cn } from '@/shared/lib/cn';
 
 interface ChatSendButtonProps {
